requirements.txt
requests
beautifulsoup4
pyyaml
pandas
python-dotenv
Flask